refactor: isolate torch-free Hugging Face management
